+++ src/pages/Bookshelf/index.ts
@@ -1,0 +1,2 @@
+export { default as BookshelfCubes } from '@/pages/Bookshelf/BookshelfCubes.vue'
+export { default as BookshelfHeader } from '@/pages/Bookshelf/BookshelfHeader.vue'